perf: move PerfReader into package and rework API
Move the PerfReader into a dedicated package, and remove
the "Perf" prefix from exported names.

Make the API more similar to the kernel interface by
renaming PerfSample to Record, and including a
counter of lost samples there. This allows users to
reconstruct at which point a sample was lost.

Also switch to a synchronous API, instead of writing
to a channel. I think this is nicer to use, and also
allows us to get rid of FlushAndClose.

Finally, clone the map passed to Reader.
This simplifies the API and makes the implementation
more robust, since the user can't break the Reader
by accidentally closing the perf_event array.

Fixes #3